+++ plugins/rest-plugin/src/main/resources/META-INF/aop.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<aspectj>
+    <aspects>
+        <aspect name="ru.sbtqa.tag.pagefactory.api.aspects.JUnitWebStepAspect"/>
+    </aspects>
+    <weaver options="-Xset:weaveJavaxPackages=true">
+        <include within="ru.sbtqa.tag..*"/>
+    </weaver>
+</aspectj>
